actionpacks: update bundled yaml to work with GNU tooling
Initial testing of the bundled yaml was done with a BSD system.
Minor inconsistencies between tool behaviour in BSD and GNU systems
causes issues as sometimes GNU tooling appends './' to folders and files.

Update bundled actions to allow it to be used as a reference in both types
of systems

Change-Id: I1922935cddd99dcb27e931424ed816d59b147f5b