Connector/grpc_client: Add ability to pass certs/key/ca as value
Currently the CA, key and certs passed to the grpc_client need to be
paths to a file.

Add the ability to specify the auth credentials' contents directly,
instead of their file locations.

This allows for connector instantiation to occur without needing
to write in-memory information to files before creating the connector

Closes=Bug721603

Change-Id: Ib1fd2585257aaed1b8af9424829133fc10a9a991