cloudvision/cvlib: fill host field when sending request to device
Method `runDeviceCmds` from the context object works as expected on
normal configuration. However configurations that have the
`--eapiOverHttps` variant activated don't support the requests sent by
the current `runDeviceCmds`.

The device interaction service expects the field `host` to be filled
when receiving a request on a configuration with the variant
`--eapiOverHttps` activated. It can be filled along with the `deviceID`
field. The device interaction will use the relevant field depending on
the current configuration to correctly identify the requested device.

Change-Id: I00991a1fbb3dd06422484640090840b03bf7ea94