cvlib/exceptions Batch InputErrors via InputErrorException
Create InputErrorException, an exception which batches InputErrors to
support error highlighting on multiple input fields in a single
exception.

Change InputError to a struct rather than a TemplateException in its
own right. InputError still contains all the data necessary to
identify and locate a Studio input that is in error.

Rename the 'inputSpecifiers' param to 'members' to match the naming
convention used in Studios code and related documentation.

Change-Id: I4ce16a284a86192646d7544bd99539ffbd19cde0