cvlib: Remove runDeviceCommands reliance on deprecated fields
Ctx.changecontrol is a deprecated field and should not be relied
upon for checking if the function is callable. Action (its replacement)
should be used instead.

Furthermore, studio autofill actions are not change controls, but may
require to run device commands to fetch such information as the
hostname.
Allow commands to be run in such contexts

Change-Id: I0dc93e50056c81f22fd2407dfd59f8eacb1fa005