cvlib: Add helper for accessing workspace aware apis
For Studio APIs, the state for a particular workspace can
be difficult to determine. A state for a particular workspace
only exists if an update has occurred for that workspace.
State may exist in mainline, or the configuration change in
the workspace may have explicitly deleted the state.

Add a helper to do all of the fallback checking until
the rAPIs return consolidated state, tracked in Bug795994

Change-Id: Ib2400259b6ea8a45577f85dccfa98e030cc5186c