cvlib/studio: Add fallback to mainline for studio input getter
Currently an annoying use case is where studio state exists in mainline
but not the workspace, and running an autofill or the like in a fresh
workspace will throw not found rpc errors. Add in a fallback to mainline
to work around that that will build up the input state. The getter will
also check to see that configs have been written that delete the input
if it needs to fall back to mainline

This change also modifies the getter to use a GetAll to handle large,
fragmented inputs, and adds the logic required to splice these together

Closes=Bug879853

Change-Id: I445bdcb504a761364d79bdffeba21ba7fda753c7