custom data for studio
The studio custom data is data generated for every build in a workspace by a studio. One studio's custom data can be read from a template by other studios from the same build or
from mainline. Dependencies between the studios must also be described in order to preserve the sequence in which they are constructed to read the latest data.

Change-Id: Ide5bb12df6c5ce778e635f4f001eeeef7a6ce819