workflows: Have sphinx install cloudvision
Currently the sphinx build is filing due to the following issue when
attempting to build
importlib.metadata.PackageNotFoundError: No package metadata was found for cloudvision

This has started to occur since the repo was moved to use a .toml
file for installation, which moved the version from being a hardcoded
string to using loading it from the importlib metadata.

The sphinx build currently does not install cloudvision, which is
causing the lookup to fail. Alter the sphinx setup to install it

Change-Id: I20d0ff07d57dabc6afef03f91ecc01c1bbeeb1a1