workflows/sphinx: Omit doctrees from published documentation
Large doctrees files (>100Mb) are causing the sphinx deploy to fail.
They are not required to display the doc correctly, and we only need
the html for that.

Omit them from the deployment of docs by adding the .doctrees folder
to the list of excluded assests in the actions-gh-pages plugin

Change-Id: I1baf6d3fa0d9d7c6c8c9aaf343e02107528e8635