repo: Update build directives to fix warnings
A number of errors are occurring in the build process due to using
old setup.py directives, and the conversion to using the toml
file has left a number of directories reporting that they would be
ignored as the namespace resolution was not being done.

Change to use a narrow namespace resolution to only include the
desired directories with the distribution to fix these errors.
This also required some inti.py file additions to ensure some
directories are included

Change-Id: Icad43d1a9bc85feabe13b846926d1b237cb9df11