cloudvision/cvlib: Do not specify path in inputs get all equality filter
Studio input getter helper methods were added in v1.7 of the library.
These initially used a GetOne approach, but were updated in v1.15 to use
a GetAll to allow for inputs over the grpc message limits to be handled,
as the Studio inputs API would break these up across multiple requests.

However, the original key from the first GetOne approach was transitioned
to be used as the partial equality filter of the GetAll. This turned out
to be problematic as that original key had a path specified as the root [].
Once the inputs crosses the grpc message limits and is broken up, the path
that the update keys contains corresponds to the subpath of the entire
inputs tree that the update came from. As those update keys no longer have
a path of [], this leads to the partial equality filter to start dropping
these updates. This causes the helper method to falsely raise an exception
that no inputs exist for the request when they do

Change-Id: I6ff6ac8ac03e52f7ab30e76d85b56d1b3dcac6e2