cvlib: Swap out deprecated crypt lib for pyavd-utils sha512 impl
Currently SHA-512-crypt is required by callers of the util, but both
crypt and passlib which offer these hashing functionalities are either
not maintained/deprecated.
pyavd-utils has some custom functionality that offers the same hashing,
using rust under the hood.
The library is actively maintained by arista so add it as a dep and
swap out the deprecated crypt lib to use the crypt hashing it offers
for the helper function.

Change-Id: I40daf03ac32bb3ec348e66c4c4c75a571cfc200e